refactored, moved Alerts within Header, moved IRMBanner inside Header
