+++ grafana-plugin/src/pages/incidents/Incidents.tsx
@@ -1,15 +1,6 @@
 import React, { ReactElement, SyntheticEvent } from 'react';
 
-import {
-  Alert as UIAlert,
-  AlertVariant,
-  Button,
-  HorizontalGroup,
-  Icon,
-  LoadingPlaceholder,
-  Tooltip,
-  VerticalGroup,
-} from '@grafana/ui';
+import { Button, HorizontalGroup, Icon, LoadingPlaceholder, Tooltip, VerticalGroup } from '@grafana/ui';
 import cn from 'classnames/bind';
 import { get } from 'lodash-es';
 import { observer } from 'mobx-react';
@@ -30,13 +21,7 @@ import { IncidentsFiltersType } from 'containers/IncidentsFilters/IncidentFilter
 import RemoteFilters from 'containers/RemoteFilters/RemoteFilters';
 import TeamName from 'containers/TeamName/TeamName';
 import { WithPermissionControlTooltip } from 'containers/WithPermissionControl/WithPermissionControlTooltip';
-import {
-  Alert,
-  Alert as AlertType,
-  AlertAction,
-  IncidentStatus,
-  IRMPlanStatus,
-} from 'models/alertgroup/alertgroup.types';
+import { Alert, Alert as AlertType, AlertAction, IncidentStatus } from 'models/alertgroup/alertgroup.types';
 import { renderRelatedUsers } from 'pages/incident/Incident.helpers';
 import { PageProps, WithStoreProps } from 'state/types';
 import { withMobXProviderContext } from 'state/withStore';
@@ -138,8 +123,6 @@ class Incidents extends React.Component<IncidentsPageProps, IncidentsPageState> 
 
     return (
       <>
-        {this.renderIRMPlanAlertMaybe()}
-
         <div className={cx('root')}>
           <div className={cx('title')}>
             <HorizontalGroup justify="space-between">
@@ -165,47 +148,6 @@ class Incidents extends React.Component<IncidentsPageProps, IncidentsPageState> 
           />
         )}
       </>
-    );
-  }
-
-  renderIRMPlanAlertMaybe() {
-    const {
-      store,
-      store: {
-        alertGroupStore: { irmPlan },
-      },
-    } = this.props;
-
-    if (store.isOpenSource) {
-      return null;
-    }
-    if (irmPlan.limits.isIrmPro || irmPlan.limits.status === IRMPlanStatus.WithinLimits) {
-      return null;
-    }
-
-    const statusSeverity: { [key: string]: AlertVariant } = {
-      [IRMPlanStatus.WithinLimits]: 'success',
-      [IRMPlanStatus.NearLimit]: 'warning',
-      [IRMPlanStatus.AtLimit]: 'error',
-    };
-
-    return (
-      <UIAlert
-        title={
-          (
-            <HorizontalGroup justify={'space-between'}>
-              <Text type={'secondary'}>
-                <div dangerouslySetInnerHTML={{ __html: irmPlan.limits.reasonHTML }} />
-              </Text>
-              <Button variant={'secondary'} onClick={() => window.open(irmPlan.limits.upgradeURL, '_blank')}>
-                Upgrade to Pro
-              </Button>
-            </HorizontalGroup>
-          ) as any
-        }
-        severity={statusSeverity[irmPlan.limits.status]}
-        buttonContent={undefined}
-      />
     );
   }
